refactor: extract TelegramService god class into focused handlers (#251)
Break 804-line TelegramService into thin orchestrator (496 lines) plus
4 focused handler classes following the existing composition pattern:

- OperationStateManager: per-queue-item locks and cancel flags
- TelegramUserManager: user creation, profile sync, membership cache
- TelegramMembershipHandler: bot added/removed from groups, onboarding
- TelegramLifecycleHandler: startup/shutdown admin notifications

Also consolidates duplicate _escape_markdown into telegram_utils.py as
the canonical location (was duplicated in telegram_service.py and
telegram_notification.py).

TelegramService keeps thin delegate methods preserving the public API so
existing handlers continue working through self.service.* unchanged.

Closes #251
